+++ rtic-macros/src/codegen/bindings.rs
@@ -2,7 +2,8 @@
     feature = "cortex-m-source-masking",
     feature = "cortex-m-basepri",
     feature = "test-template",
-    feature = "riscv-esp32c3"
+    feature = "riscv-esp32c3",
+    feature = "riscv-slic",
 )))]
 compile_error!("No backend selected");
 
@@ -23,3 +24,9 @@ pub use esp32c3::*;
 
 #[cfg(feature = "riscv-esp32c3")]
 mod esp32c3;
+
+#[cfg(feature = "riscv-slic")]
+pub use riscv_slic::*;
+
+#[cfg(feature = "riscv-slic")]
+mod riscv_slic;