+++ rtic/CHANGELOG.md
@@ -17,6 +17,7 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ### Added
 
+- Unstable support for RISC-V targets compatible with `riscv-slic`
 - Unstable support for ESP32-C3
 
 ### Fixed